Create utils-package and move classes there
