+++ src/main/java/services/interfaces/SchemaParser.java
@@ -2,7 +2,7 @@ package services.interfaces;
 
 import java.util.List;
 
-import services.dataclasses.OntologyConcept;
+import services.utils.OntologyConcept;
 
 public interface SchemaParser {
   public List<OntologyConcept> parse(String filePath) throws Exception;
