+++ src/main/java/services/utils/Unzipper.java
@@ -1,4 +1,4 @@
-package services.parsers;
+package services.utils;
 
 import java.io.File;
 
@@ -6,11 +6,11 @@ import org.apache.commons.io.FilenameUtils;
 
 import net.lingala.zip4j.core.ZipFile;
 
-public class ZipParser {
+public class Unzipper {
   /**
    * Unzips and returns the location of the unzipped folder.
    */
-  public static String Unzip(String inputPath, String outputPath) throws Exception{
+  public static String unzip(String inputPath, String outputPath) throws Exception{
     ZipFile zipFile = new ZipFile(inputPath);
     zipFile.extractAll(outputPath);
     File file = new File(outputPath);
@@ -27,9 +27,9 @@ public class ZipParser {
   /**
    * Unzips and returns the location of the unzipped folder.
    */
-  public static String Unzip(String path) throws Exception{
+  public static String unzip(String path) throws Exception{
     String outputPath = FilenameUtils.removeExtension(path);
-    return Unzip(path, outputPath);
+    return unzip(path, outputPath);
   }
 
 }
